Kooppagina enzo

+++ app/Http/Controllers/GebouwenController.php
@@ -15,6 +15,8 @@ class GebouwenController extends Controller
      */
     public function index()
     {
+        if(!auth()->user()->role == 'admin') # Deze methode is slecht en sloom, maar ik weet nu even niks beters.
+            return redirect(route('login'));
         return view('gebouwen');
     }
 
@@ -25,6 +27,8 @@ class GebouwenController extends Controller
      */
     public function create()
     {
+        if(!auth()->user()->role == 'admin') # Deze methode is slecht en sloom, maar ik weet nu even niks beters.
+            return redirect(route('login'));
         return view('gebouwen.create');
     }
 
@@ -36,6 +40,8 @@ class GebouwenController extends Controller
      */
     public function store(Request $request)
     {
+        if(!auth()->user()->role == 'admin') # Deze methode is slecht en sloom, maar ik weet nu even niks beters.
+            return redirect(route('login'));
         $gebouw = new Gebouw;
         $gebouw->naam = $request->gebouw_naam;
         $gebouw->prijs = $request->gebouw_prijs;
@@ -44,7 +50,7 @@ class GebouwenController extends Controller
         else 
             $gebouw->uuid = $request->gebouw_uuid;
         $gebouw->save();
-        return redirect(route('dashboard.overview'));
+        return redirect(route('dashboard.overview'))->with(['success'=>'Het gebouw ' . $gebouw->naam . ' is aangemaakt.']);
     }
 
     /**
@@ -82,6 +88,8 @@ class GebouwenController extends Controller
     }
 
     public function delete() {
+        if(!auth()->user()->role == 'admin') # Deze methode is slecht en sloom, maar ik weet nu even niks beters.
+            return redirect(route('login'));
         $gebouwen = Gebouw::get();
         return view('gebouwen.delete')->with(compact('gebouwen'));
     }
@@ -94,8 +102,10 @@ class GebouwenController extends Controller
      */
     public function destroy(Request $request)
     {
+        if(!auth()->user()->role == 'admin') # Deze methode is slecht en sloom, maar ik weet nu even niks beters.
+            return redirect(route('login'));
         Gebouw::destroy($request->gebouw_id);
-        return redirect(route('dashboard.overview'));
+        return redirect(route('dashboard.overview'))->with(['success'=>'Het gebouw is verwijderd.']);
     }
 
     /**
@@ -104,9 +114,11 @@ class GebouwenController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function buy($id)
+    public function buy($uuid)
     {
-        //
+        $gebouw = Gebouw::where('uuid', $uuid)->first();
+        if(!$gebouw) return redirect(route('dashboard.overview'))->with(['error'=>'Dat gebouw kan je niet kopen']);
+        return view('buy')->with(compact('gebouw'));
     }
 
 }
