+++ app/Http/Controllers/UsersController.php
@@ -3,10 +3,11 @@
 namespace App\Http\Controllers;
 
 use App\Models\User;
+use App\Models\Gebouw;
 use Illuminate\Support\Str;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
-use App\Models\Gebouw;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Hash;
 
 class UsersController extends Controller
@@ -21,6 +22,11 @@ class UsersController extends Controller
         //
     }
 
+    private function adminCheck() {
+        if(!auth()->user()->role == 'admin') # Deze methode is slecht en sloom, maar ik weet nu even niks beters.
+        return redirect(route('login'));
+    }
+
     /**
      * Show the form for creating a new resource.
      *
@@ -28,6 +34,8 @@ class UsersController extends Controller
      */
     public function create()
     {
+        if(!auth()->user()->role == 'admin') # Deze methode is slecht en sloom, maar ik weet nu even niks beters.
+            return redirect(route('login'));
         return view('users.create');
     }
 
@@ -39,13 +47,14 @@ class UsersController extends Controller
      */
     public function store(Request $request)
     {
+        if(!auth()->user()->role == 'admin') # Deze methode is slecht en sloom, maar ik weet nu even niks beters.
+            return redirect(route('login'));
         $user = new User;
         $user->name = $request->user_naam;
-        $user->email = $request->user_email;
         $user->password = Hash::make($request->user_wachtwoord);
         $user->role = "gebruiker";
         $user->save();
-        return redirect(route('dashboard.management'));
+        return redirect(route('dashboard.overview'))->with(['success'=>'Het team genaamd ' . $request->user_naam . ' is aangemaakt.']);
     }
 
     /**
@@ -95,18 +104,24 @@ class UsersController extends Controller
      */
     public function destroy(Request $request)
     {
+        if(!auth()->user()->role == 'admin') # Deze methode is slecht en sloom, maar ik weet nu even niks beters.
+            return redirect(route('login'));
         $id = $request->user_id;
         if($id == 1) return redirect(route('dashboard.overview'));
 
         # Zoek gebouwen en onteigen ze.
-        $this->detachFromModels($id);
+        // $this->detachFromModels($id);
+        $gebouwen = Gebouw::where('user_id', $id);
+        $gebouwen->update(['user_id'=>null]);
 
         User::destroy($id);
-        return redirect(route('dashboard.overview'));
+        return redirect(route('dashboard.overview'))->with(['success'=>'Het team is verwijderd.']);
     }
 
     public function delete()
     {
+        if(!auth()->user()->role == 'admin') # Deze methode is slecht en sloom, maar ik weet nu even niks beters.
+            return redirect(route('login'));
         $users = User::get();
         return view('users.delete')->with(compact('users'));
     }
